Allow default_permissions and allow_other to be disabled.
Ever since 0528a7c these options have been always-on.  Now they will
actually be defaults: if the user supplies nodefault_permissions or
noallow_other on the command line, then default_permissions or
allow_other will not be set, respectively.

This commit does not change the behavior of local or noappledouble on
OSX, because I don't have any Apple systems to test it on.

Fixes #95